improved data transfer, added homepage, filelist
server will only allow http requests to read from the files in
"validFiles.ini" in the root of the repository directory. every line is a
file in /serverdata that will be allowed. i.e. if there's a line
"/index.html", the server will allow http requests to get
"/serverdata/index.html". the default homepage is "/index.html".